requirements.txt
# Databricks SDK and CLI
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# Streamlit for web app
streamlit>=1.28.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Optional: Additional useful packages
requests>=2.31.0
